Huge refactor of buildscript
- Store versions in version catalogs
- Move any duplicate logic to convention plugins
- Refactor modules
  - Refactor main project -> api subproject
  - Refactor compose subproject -> gui subproject
- Add axion-release plugin for versioning
- Add typesafe project accessors for depending on subprojects
- Add dokka for documentation generation
- Clean up & refactor publishing
