+++ api/src/dbGen/kotlin/dev/kosmx/needle/dbGen/db/Skyrage.kt
