+++ api/src/dbGen/kotlin/dev/kosmx/needle/dbGen/dsl/TypeEntry.kt
