+++ api/src/main/kotlin/dev/kosmx/needle/Static.kt
