+++ api/src/main/kotlin/dev/kosmx/needle/core/AssetChecker.kt
