+++ api/src/main/kotlin/dev/kosmx/needle/core/ClassChecker.kt
