+++ api/src/main/kotlin/dev/kosmx/needle/core/JarCheckResult.kt
