+++ api/src/main/kotlin/dev/kosmx/needle/database/hardCodedDetectors/HardCodedDetectors.kt
