+++ api/src/test/kotlin/dev/kosmx/jarchecker/lib/MkConfig.kt
