+++ dokka/includes/Module.md
@@ -1,0 +1,37 @@
+# Module jneedle-api
+
+Yet another jar malware detection tool
+
+## Features
+
+- TODO
+
+## Including
+
+You can include jNeedle in your project by adding the following:
+
+### Maven
+
+```xml
+<dependency>
+  <groupId>${project.group}</groupId>
+  <artifactId>${project.module}</artifactId>
+  <version>${project.version}</version>
+</dependency>
+```
+
+### Gradle Groovy DSL
+
+```groovy
+implementation '${project.group}:${project.module}:${project.version}'
+```
+
+### Gradle Kotlin DSL
+
+```kotlin
+implementation("${project.group}:${project.module}:${project.version}")
+```
+
+## Examples
+
+TODO
