+++ settings.gradle.kts
@@ -9,4 +9,9 @@ plugins {
 }
 
 rootProject.name = "jneedle"
-include(":compose")
+include(":api")
+//include(":cli") // TODO: 2023-06-12 Move cli shit into cli
+include(":gui")
+
+enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
+
