--- src/main/kotlin/dev/kosmx/needle/database/AssetMatch.kt
@@ -1,9 +1,0 @@
-package dev.kosmx.needle.database
-
-import org.objectweb.asm.tree.AbstractInsnNode
-import software.coley.llzip.format.model.LocalFileHeader
-import java.util.jar.JarEntry
-
-interface AssetMatch : Match {
-    fun match(data: Lazy<ByteArray>, jarEntry: LocalFileHeader): Boolean
-}
